Move color palette + should_use_color into leaf module imbue.mngr.colors
Fixes an import-linter layer violation: `imbue.mngr.errors` is below
`imbue.mngr.utils` in the "mngr layers contract", so errors.py importing
the color helpers from utils/logging.py (added for MngrError.show) was a
lower layer reaching up into a higher one.

Extract the ANSI palette and should_use_color into a new stdlib-only leaf
module `imbue.mngr.colors` (not part of the layers contract, like the
existing `imbue.mngr.plugin_catalog` that errors.py already imports). Both
errors.py and utils/logging.py now import from it -- one source of truth,
no layer violation, no circular dependency. Moves the should_use_color
tests to colors_test.py accordingly.
